requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
#mysqlclient
#psycopg2-binary
psycopg2
#sqlparse
django-rest-swagger
uWSGI>=2.0.18,<2.1
intuit-oauth